requirements.txt
tensorflow==2.18.0
matplotlib==3.9.2
numpy==1.26.3
streamlit==1.39.0
Pillow
streamlit-drawable-canvas
opencv-python